requirements.txt
radioactivedecay
pandas
matplotlib
Pillow